requirements.txt
# Core Dependencies
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0

# AI/LLM - Google Gemini
google-generativeai==0.3.2

# Configuration
pyyaml==6.0.1

# Data Processing
numpy==1.26.3
pandas==2.1.4

# Performance & Monitoring
psutil==5.9.7

# Testing (Optional but recommended)
pytest==7.4.4
pytest-asyncio==0.23.3

# HTML Parsing Alternatives (backup)
html5lib==1.1

# Utilities
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1

# Logging
loguru==0.7.2

# JSON Schema Validation
jsonschema==4.20.0

# Date/Time utilities
python-dateutil==2.8.2